LayoutTests/webgpu/render-command-encoding.html

<!DOCTYPE html><!-- webkit-test-runner [ experimental:WebGPUEnabled=true ] -->
<meta charset=utf-8>
<title>Get the WebGPUDevice and ask for the WebGPUQueue</title>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="js/webgpu-functions.js"></script>
<script>
const shaders = `
#include <metal_stdlib>
    
using namespace metal;

struct Vertex
{
    float4 position [[position]];
};

vertex Vertex vertex_main()
{
    return Vertex { float4(0, 0, 0, 1) };
}

fragment float4 fragment_main(Vertex vertexIn [[stage_in]])
{
    return vertexIn.position;
}
`
promise_test(async () => {
    const device = await getBasicDevice();

    const canvas = document.createElement("canvas");
    const swapChain = createBasicSwapChain(canvas, device);
    const shaderModule = device.createShaderModule({ code: shaders });
    const pipeline = createBasicPipeline(shaderModule, device);

    const commandBuffer = device.createCommandBuffer();
    assert_true(commandBuffer instanceof WebGPUCommandBuffer, "createCommandBuffer returned a WebGPUCommandBuffer");

    const encoder = beginBasicRenderPass(swapChain, commandBuffer);
    assert_true(encoder instanceof WebGPURenderPassEncoder, "beginRenderPass() returned a WebGPURenderPassEncoder");

    encoder.setPipeline(pipeline);
    const commandBufferEnd = encoder.endPass();
    assert_true(commandBufferEnd instanceof WebGPUCommandBuffer, "endPass() returned a WebGPUCommandBuffer");
}, "WebGPURenderPassEncoder created and successfully ended");

</script>
</body>
